% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRefs.R
\name{ABItgt}
\alias{ABItgt}
\title{ABItgt()
Computes ABI for target F, e.g. ABImsy (Griffith et al. 2023)}
\usage{
ABItgt(stock, ftgt = 0.2, thresh = 0.9)
}
\arguments{
\item{stock}{object of class FLStock}

\item{ftgt}{target F at equilibrium, e.g. Fmsy}

\item{thresh}{quantile ageref treshold, default 0.9}
}
\value{
*FLQuant*
}
\description{
ABItgt()
Computes ABI for target F, e.g. ABImsy (Griffith et al. 2023)
}
\examples{
data(ple4)
ABImsy = ABItgt(ple4,ftgt=0.22,thresh=0.9)
plot(ABImsy)+ylim(0,2)+
 geom_hline(yintercept = c(0.8,1),col=c(2,1),linetype=c(2,1))+ylab(expression(ABI[MSY]))
}
