% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lensim.R
\name{alk.sample}
\alias{alk.sample}
\title{Sample annual age-length keys by length class}
\usage{
alk.sample(lfds, alks, nbin = 20, n.sample = 1)
}
\arguments{
\item{lfds}{An `FLQuant` containing length-frequency data by year and
iteration.}

\item{alks}{An `FLPars` object containing annual age-length keys, typically
returned by [ALKs()].}

\item{nbin}{Maximum number of age observations sampled per length class.
Defaults to `20`.}

\item{n.sample}{Optional total length-sample size used to rescale `lfds`.
A value of `1` uses the supplied frequencies without rescaling.}
}
\value{
An `FLPars` object containing sampled annual age-length keys.
}
\description{
Generates multinomial age samples within length classes from annual
age-length keys. The number sampled in each length class is limited by
`nbin` and by the corresponding length-frequency sample.
}
\examples{
\dontrun{
sampled_alks <- alk.sample(lfds, alks, nbin = 20)
}

}
