% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lensim.R
\name{applyALK}
\alias{applyALK}
\title{applyALK function to length to age}
\usage{
applyALK(lfds, alks)
}
\arguments{
\item{alks}{*FLPars* annual ALKs}

\item{lfd}{*FLQuant* with numbers at length}
}
\value{
FLQuant for numbers at age
}
\description{
applyALK function to length to age
}
