% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.R
\name{applyAR}
\alias{applyAR}
\title{applyAR()
Function to apply ICES Advice rule for F < MSY Btrigger}
\usage{
applyAR(b, btrigger, fmsy, bmin = 0, fmin = 0.001)
}
\arguments{
\item{b}{biomass}

\item{btrigger}{if b below btrigger F is reduced}

\item{fmsy}{Fmsy target}

\item{bmin}{Option for de facto fishing closure (e.g. bmin = blim)}

\item{fmin}{F below bmin}
}
\value{
F advice
}
\description{
applyAR()
Function to apply ICES Advice rule for F < MSY Btrigger
}
