% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blag}
\alias{blag}
\title{blag()}
\usage{
blag(mvn, verbose = TRUE)
}
\arguments{
\item{mvn}{}
}
\value{
output list of quant posteriors and mle's
}
\description{
function to assign B[y+1] to B[y]. Warning correlation structure of B[y+1] and F[y] is meaningless
}
\author{
Henning Winker (GFCM)
}
