% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lensim.R
\name{catch_gear}
\alias{catch_gear}
\title{Catch by gear and year}
\usage{
catch_gear(catch_n, wt = NULL, type = c("weight", "numbers"))
}
\arguments{
\item{catch_n}{`FLQuants` of catch numbers-at-age by gear
(from [catch_n_gear()]).}

\item{wt}{`FLStock` (its `catch.wt` is used) or an `FLQuant` of
weight-at-age. Ignored for `type = "numbers"`.}

\item{type}{`"weight"` (default) or `"numbers"`.}
}
\value{
`FLQuants` of annual catch by gear (age summed).
}
\description{
Catch by gear and year
}
