% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRefs.R
\name{computeFbrp}
\alias{computeFbrp}
\title{computeFbrp()
Computes biological reference points corresponding to the proxy Fbrp}
\usage{
computeFbrp(
  stock,
  sr = "missing",
  proxy = NULL,
  x = NULL,
  blim = 0.1,
  type = c("b0", "btrg", "value"),
  btri = "missing",
  bpa = "missing",
  verbose = T,
  fmax = 5
)
}
\arguments{
\item{stock}{object of class FLStock}

\item{sr}{stock recruitment model of class FLSR}

\item{x}{basis in percent for sprx and bx, e.g. 40 for spr40}

\item{blim}{values < 1 are taken as fraction to B0 and blim > 1 as absolute values unless specified otherwise}

\item{type}{type of blim input, values < 1 are
\itemize{
  \item "b0" fraction to B0  
  \item "btrg" fraction to Btarget (first occurring in proxy)    
  \item "value" absolute value
}}

\item{btri}{Btrigger can specified as absolute value}

\item{bpa}{Bpa can specified as absolute value}

\item{verbose}{}

\item{fmax}{maximum Flim = max(Flim,fmax*Fbrp)}

\item{proxies}{choice of Fmsy proxies (combinations permitted) 
\itemize{
  \item "sprx"  spawning potential ratio spr/spr0 with basis x 
  \item "bx" SSB as fraction xSSB0
  \item "f0.1" 10% slope of yield-per-recruit curve
  \item "msy"  maximum surplus production (not defined for segreg)
}}
}
\value{
brp object of class FLBRP with computed Fbrp reference points
}
\description{
computeFbrp()
Computes biological reference points corresponding to the proxy Fbrp
}
\examples{
data(ple4)
srr = srrTMB(as.FLSR(ple4,model=rickerSV),spr0=spr0y(ple4))
brp = computeFbrp(stock=ple4,sr=srr,proxy=c("sprx","f0.1"),blim=0.1,type="b0")
ploteq(brp,obs=TRUE,refpts="msy")
}
