% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRefs.R
\name{computeFmmy}
\alias{computeFmmy}
\title{computeFmmy()
Uses opt.bisect to derive the F at Maximum Median Yield from stochastic simulations}
\usage{
computeFmmy(
  brp,
  sigmaR = 0.5,
  rho = 0,
  nyears = 100,
  iters = 250,
  yrs.eval = NULL,
  range = "missing",
  tol = 0.01,
  maxit = 15,
  verbose = TRUE
)
}
\arguments{
\item{brp}{output object from computeFbrp() of class FLBRP}

\item{sigmaR}{lognormal recruitment standard deviation}

\item{rho}{AR1 recruitment autocorrelation coefficient}

\item{nyears}{number of simulation years}

\item{iters}{number simulation iterations}

\item{yrs.eval}{last years to be used evaluation period, default nyears/2}

\item{range}{range of Fbar value to be evaluated}

\item{tol}{tolerance}

\item{maxit}{number of steps}

\item{verbose}{cat comments}
}
\value{
list of  FLPar, FLStock and FLBRP objects
}
\description{
computeFmmy()
Uses opt.bisect to derive the F at Maximum Median Yield from stochastic simulations
}
\examples{
data(ple4)
srr = srrTMB(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4),plim=0.15)
brp=computeFbrp(ple4,st,proxy = c("msy","sprx"),x=35,blim=params(st)[[2]])
ploteq(brp)
# DEFINE MMY statistic
statistic <- list(MMY=list(~apply(L,1,median), name="MMY",
  desc="ICES Maximum Median Yield"))
# CALL bisect over 100 years, Fmmy calculated over last 50.
fmmy <- opt.bisect(stock, sr=srr, deviances=devs, metrics=list(L=landings), 
statistic=statistic, years=2018:2118,
pyears=2069:2118, tune=list(fbar=c(0.01, 0.2)))
# fmmy
mean(fbar(fmmy)[,ac(2069:2118)]) 


}
