% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_la.R
\name{f_share_gear}
\alias{f_share_gear}
\title{Time-varying share of total F-at-age by gear}
\usage{
f_share_gear(f_age)
}
\arguments{
\item{f_age}{`FLQuants` of gear F-at-age, e.g. from [f_age_gear()].}
}
\value{
`FLQuants` of shares (0 where total F is 0).
}
\description{
\deqn{\phi_{g,a,y}=F_{g,a,y}\Big/\sum_h F_{h,a,y}}
General form of `om_gears[[g]]$f_share`, needed when gear Fbar
trajectories are not proportional (e.g. an MP that manages one gear only).
}
