% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRefs.R
\name{Fbrp}
\alias{Fbrp}
\title{Fbrp()
Extract Fbrp based reference points from output of computeFbrp}
\usage{
Fbrp(brp)
}
\arguments{
\item{brp}{input of class FLBRP from ComputeFbrp}
}
\value{
FLPar object with computed Fbrp reference points
}
\description{
Fbrp()
Extract Fbrp based reference points from output of computeFbrp
}
