% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRefs.R
\name{Fe40}
\alias{Fe40}
\title{Fe40()
Patterson estimator for Fmsy}
\usage{
Fe40(stock, nyears = 3)
}
\arguments{
\item{stock}{input of class FLStock}

\item{nyears}{number of years to average}
}
\value{
value
}
\description{
Fe40()
Patterson estimator for Fmsy
}
