% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{flr2stars}
\alias{flr2stars}
\title{jabba2stars()}
\usage{
flr2stars(object, quantiles = c(0.025, 0.975))
}
\arguments{
\item{object}{of class FLStockR}

\item{quantities}{default is 95CIs as c(0.025,0.975)}
}
\value{
STARS list with $timeseris and $refpts
}
\description{
jabba2stars()
}
