% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLRef.R
\docType{package}
\name{FLRef}
\alias{FLRef-package}
\alias{FLRef}
\title{FLRef: Reference point computation for advice rules}
\description{
Blah
}
\author{
\strong{Maintainer}: Henning Winker \email{henning.winker@gmail.com}

}
\keyword{internal}
