% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRefs.R
\name{Fp05}
\alias{Fp05}
\title{Calculates the Fbar value giving a maximum probability of ssb being below Blim of 5 percent}
\usage{
Fp05(
  object,
  iters = "missing",
  range = "missing",
  tol = 0.001,
  maxit = 20,
  verbose = TRUE
)
}
\arguments{
\item{object}{output from Fsim()}

\item{iters}{Number of iterations, cannot exceed input object}

\item{range}{range of Fbar value to be evaluated}

\item{verbose}{Should progress be shown, TRUE.}
}
\value{
list
}
\description{
Calculates the Fbar value giving a maximum probability of ssb being below Blim of 5 percent
}
\examples{
data(ple4)
bh = srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=spr0y(ple4))
brp = computeFbrp(ple4,bh,proxy="bx",x=35,blim=0.1)
fsim = Fsim(brp,sigmaR=0.7,rho=0.3)
plotFsim(fsim)
fp.05 = Fp05(fsim)
plotFsim(fp.05,panels=c(2,4)) # black line is Fp0.05
getF(fp.05) 
}
