% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRefs.R
\name{Fsim}
\alias{Fsim}
\title{Fsim()
Simulates stochastic stock dynamics under under constant Fbrp}
\usage{
Fsim(
  brp,
  sigmaR = 0.5,
  rho = 0,
  nyears = 100,
  iters = 250,
  yrs.eval = NULL,
  verbose = TRUE
)
}
\arguments{
\item{brp}{output object from computeFbrp() of class FLBRP}

\item{sigmaR}{lognormal recruitment standard deviation}

\item{rho}{AR1 recruitment autocorrelation coefficient}

\item{nyears}{number of simulation years}

\item{iters}{number simulation iterations}

\item{yrs.eval}{last years to be used evaluation period, default nyears/2}

\item{verbose}{cat comments}
}
\value{
list of  FLPar, FLStock and FLBRP objects
}
\description{
Fsim()
Simulates stochastic stock dynamics under under constant Fbrp
}
\examples{
data(ple4)
hs = srrTMB(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4),lplim=0.05,uplim=0.25)
blim = params(hs)[[2]]
brp = computeFbrp(ple4,hs,proxy=c("sprx","f0.1","msy"),x=40,blim=blim)
ploteq(brp)
fsim = Fsim(brp,sigmaR=0.7,rho=0.3)
plotFsim(fsim)
plotFsim(fsim,panels=2)
}
