% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{fudc}
\alias{fudc}
\title{generates an up-down-constant F-pattern}
\usage{
fudc(
  object,
  fref = 0.2,
  fhi = 2.5,
  flo = 0.8,
  sigmaF = 0.2,
  breaks = c(0.5, 0.75)
)
}
\arguments{
\item{object}{An *FLStock*}

\item{fref}{reference denominator for fbar}

\item{fhi}{factor for high F as fhi = fbar/fref}

\item{flo}{factor for low F as flo = fbar/fref}

\item{sigmaF}{variation on fbar}

\item{breaks}{relative location of directional change}
}
\value{
FLQuant
}
\description{
generates an up-down-constant F-pattern
}
\examples{
data(ple4)
sr <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
brp = computeFbrp(ple4,sr,proxy="msy")
fmsy = Fbrp(brp)["Fmsy"]
stki = propagate(ple4,100)
fy = fudc(ple4,fhi=2,flo=0.9,fref=fmsy,sigmaF=0)
fyi = fudc(stki,fhi=2,flo=0.9,fref=fmsy,sigmaF=0.2)
plot(fy,fyi)+ylab("F")
#Forcasting
om <- FLStockR(ffwd(stki,sr,fbar=fyi))
om@refpts = Fbrp(brp)
plotAdvice(window(om,start=1960))
}
