% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices.R
\name{fwd2ices}
\alias{fwd2ices}
\title{Function to summarise forecast results}
\usage{
fwd2ices(stock, uncertainty, eval.yrs = NULL, dB = NULL, refyr = NULL)

fwd2ices(stock, uncertainty, eval.yrs = NULL, dB = NULL, refyr = NULL)
}
\arguments{
\item{uncertainty}{*FLStocks* with list of *FLStockR* objects and iters}

\item{eval.yrs}{evaluation years of forecast}

\item{dB}{computes change in percentage biomass to refyr}

\item{refyr}{change in percentage biomass to refyr}

\item{object}{*FLStocks* with list of *FLStockR* objects}

\item{rel}{if TRUE ratios B/Btgt and F/Ftgt are shown}

\item{stk}{*FLStocks* with list of *FLStockR* objects}
}
\value{
data.frame

data.frame
}
\description{
Function to summarise forecast results

Function to summarise ICES fishing opportunities
}
