% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{fwd2stars}
\alias{fwd2stars}
\title{Function to summarise forecast results}
\usage{
fwd2stars(object, eval.yrs = NULL, rel = NULL, dB = NULL, refyr = NULL)
}
\arguments{
\item{object}{*FLStocks* with list of *FLStockR* objects}

\item{eval.yrs}{evaluation years of forecast}

\item{rel}{if TRUE ratios B/Btgt and F/Ftgt are shown}

\item{dB}{computes change in percentage biomass to refyr}
}
\value{
data.frame
}
\description{
Function to summarise forecast results
}
