% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRefs.R
\name{fwdF4B}
\alias{fwdF4B}
\title{fwdF4B ()
Bisection Function to search F for a given biomass}
\usage{
fwdF4B(
  stock,
  sr,
  btgt,
  nfy = 3,
  niy = 1,
  ival = niy,
  imet = "TAC",
  ftune = c(0, 2),
  tol = 0.001,
  verbose = TRUE,
  ssbQ = 1,
  recQ = 1
)
}
\arguments{
\item{stock}{FLStock object}

\item{sr}{stock recruitment relationship}

\item{btgt}{target biomass of the search}

\item{nfy}{number of forecast (default 3)}

\item{niy}{number of intermediate years (default 1)}

\item{ival}{intermediate year value}

\item{imet}{intermediate year metric ("TAC","F")}

\item{ftune}{tuning limits for F search c(0.1,2)}

\item{tol}{precision tolerance}

\item{verbose}{}
}
\value{
*FLStock*
}
\description{
fwdF4B ()
Bisection Function to search F for a given biomass
}
