% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{idx.sim}
\alias{idx.sim}
\title{generates FLIndex with lognormal annual and multinomial age composition observation error}
\usage{
idx.sim(
  object,
  sel = catch.sel(object),
  ages = NULL,
  ess = 200,
  sigma = 0.2,
  q = 0.01
)
}
\arguments{
\item{object}{FLStock}

\item{sel}{FLQuant with selectivity.pattern}

\item{ess}{effective sample size for age composition sample}

\item{sigma}{annual observation error for log(q)}

\item{q}{catchability coefficient for scaling}

\item{age}{define age range}
}
\value{
FLIndex
}
\description{
generates FLIndex with lognormal annual and multinomial age composition observation error
}
\examples{
data(ple4)
sel = newselex(catch.sel(ple4),FLPar(S50=1.5,S95=2.1,Smax=4.5,Dcv=1,Dmin=0.1))
ggplot(sel)+geom_line(aes(age,data))+ylab("Selectivity")+xlab("Age")
object = propagate(ple4,10)
idx = idx.sim(object,sel=sel,ess=200,sigma=0.2,q=0.01)
# Checks
ggplot(idx@sel.pattern)+geom_line(aes(age,data))+ylab("Selectivity")+xlab("Age")
ggplot(idx@index)+geom_line(aes(year,data,col=ac(iter)))+facet_wrap(~age,scales="free_y")+
theme(legend.position = "none")+ylab("Index")
}
