% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{jabba2FLStockR}
\alias{jabba2FLStockR}
\title{jabba2FLStockR()}
\usage{
jabba2FLStockR(jabba, blim = 0.3, bthr = 0.5, thin = 10, rel = FALSE)
}
\arguments{
\item{jabba}{fit from JABBA fit_jabba() or jabba$kbtrj}

\item{blim}{biomass limit reference point as fraction of Bmsy}

\item{thin}{thinnig rate of retained iters}

\item{rel}{if TRUE ratios BBmsy and FFmsy are stored}

\item{bpa}{biomass precautionary reference point as fraction of Bmsy}
}
\value{
FLStockR with refpts
}
\description{
jabba2FLStockR()
}
