% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{jabba2stars}
\alias{jabba2stars}
\title{jabba2stars()}
\usage{
jabba2stars(jabba, quantiles = c(0.025, 0.975), bfrac = 0.3)
}
\arguments{
\item{jabba}{fit from JABBA fit_jabba() or jabba$kbtrj}

\item{quantiles}{default is 95CIs as c(0.025,0.975)}

\item{bfrac}{biomass fraction of Bmsy, default 0.3Bmsy (ICES)}
}
\value{
STARS list with $timeseris and $refpts
}
\description{
jabba2stars()
}
