% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mlorenzen}
\alias{Mlorenzen}
\title{Mlorenzen}
\usage{
Mlorenzen(object, Mref = "missing", Aref = 2)
}
\arguments{
\item{object}{weight-at-age of class *FLQuant*}

\item{Mref}{reference M for scaling}

\item{Aref}{reference Age for scaling}
}
\value{
FLQuant m()
}
\description{
computes Lorenzen M with scaling option
}
\examples{
data(ple4)
Ml = Mlorenzen(stock.wt(ple4))
# Scale
Ms = Mlorenzen(stock.wt(ple4),Mref=0.2,Aref=2)
flqs = FLQuants(Lorenzen=Ml,Scaled=Ms)
}
