% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesim.R
\name{plot_pe}
\alias{plot_pe}
\title{Plot biomass process error against B/Bmsy}
\usage{
plot_pe(ps, rel = TRUE, ref = 1, line = TRUE, points = TRUE, reverse_x = FALSE)
}
\arguments{
\item{ps}{data.frame returned by pe_sampler(), with columns sigma and bbmsy.}

\item{rel}{Logical. If TRUE, both B/Bmsy and sigma are divided by
their reference values.}

\item{ref}{Integer. Reference row used for normalisation. Defaults to 1.}

\item{line}{Logical. Add line.}

\item{points}{Logical. Add points.}

\item{reverse_x}{Logical. Reverse x-axis.}
}
\value{
A ggplot object.
}
\description{
Plots sampled biomass process error against biomass relative to Bmsy.
Optionally both axes can be normalised relative to a reference row,
typically row 1, which corresponds to the Fmsy scenario in Fsim_grid().
}
