% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesim.R
\name{plot_pf}
\alias{plot_pf}
\title{plot_pf()}
\usage{
plot_pf(object, quant = c("vb", "ssb"), fmsy = NULL, rel = FALSE)
}
\arguments{
\item{object}{An *FLBRP*}

\item{quant}{choose between vb and ssb or both}

\item{fmsy}{default if Fmsy}

\item{rel}{if TRUE ratios are produced for spcurve}
}
\value{
ggplot
}
\description{
plots production functions
}
\examples{
data(ple4)
ple4 <- merge_catch(ple4)
sr <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
brp = brp(FLBRP(ple4,sr))
asem2spm(brp)[1:4]
plot_pf(brp)
plot_pf(brp,rel=TRUE)
}
