% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_la.R
\name{plot_sel_age}
\alias{plot_sel_age}
\title{Plot selectivity-at-age}
\usage{
plot_sel_age(
  object,
  age_by = 1,
  points = TRUE,
  title = NULL,
  line_width = 0.7,
  point_size = 1.5,
  colours = NULL,
  ylab = "Selectivity",
  xlab = "Age"
)
}
\arguments{
\item{object}{An `FLQuant` containing selectivity-at-age, or a list of
`FLQuant` objects. If a list is supplied, each element is plotted as a
separate curve. List names are used as legend labels.}

\item{age_by}{Numeric. Spacing of age-axis breaks. Default is `1`.}

\item{points}{Logical. Should points be added to the curves? Default is
`TRUE`.}

\item{title}{Optional plot title. Can be a character string or expression.}

\item{line_width}{Numeric. Line width passed to `ggplot2::geom_line()`.
Default is `0.7`.}

\item{point_size}{Numeric. Point size passed to `ggplot2::geom_point()`.
Default is `1.5`.}

\item{colours}{Optional named vector of colours. Names must match curve
names. If `NULL`, ggplot2 default colours are used.}

\item{ylab}{Character. Y-axis label. Default is `"Selectivity"`.}

\item{xlab}{Character. X-axis label. Default is `"Age"`.}
}
\value{
A `ggplot` object.
}
\description{
Plots one or more selectivity-at-age curves stored as `FLQuant` objects.
}
\details{
The function accepts either a single `FLQuant` or a named list of `FLQuant`
objects. It is intended for comparing joint or gear-specific
selectivity-at-age curves used in operating models.


The age dimension is extracted from the `age` dimension of each `FLQuant`.
Plus-group labels such as `"10+"` are converted to numeric values for
plotting.

If a single `FLQuant` is provided, the legend is suppressed. If a list of
`FLQuant` objects is provided, the legend is shown using the list names.
}
\examples{
\dontrun{
## Single curve
plot_sel_age(sel_joint)

## Multiple species
sel_joint <- list(
  siganus = sel_joint_siganus,
  lethrinus = sel_joint_lethrinus,
  lutjanus = sel_joint_lutjanus,
  shark = sel_joint_shark
)

plot_sel_age(sel_joint)
}

}
