% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotAR}
\alias{plotAR}
\title{plotAR
Plots the new proposed ICES advice rule}
\usage{
plotAR(
  pars,
  ftrg = 1,
  btrigger = "missing",
  bpa = "missing",
  bclose = 0,
  fmin = 0,
  obs = "missing",
  kobe = TRUE,
  alpha = 1,
  xmax = 1.2,
  ymax = 1.5,
  ylab = "missing",
  xlab = "missing",
  rel = FALSE,
  expand = TRUE,
  labels = TRUE,
  label.cex = 3,
  critical = TRUE
)
}
\arguments{
\item{pars}{FLPar object or computeFbrp() ouput 
\itemize{
  \item 1: "Fbrp"  # "F.." must first
  \item 2: "Btrg"
  \item 3: "Blim"
  \item 4: "B0"
}}

\item{ftrg}{factor to adjust Fmsy or its proxy e.g. 0.8Fmsy}

\item{btrigger}{biomass trigger below which F is linearly reduced}

\item{bpa}{biomass threshold above which biomass is considered precautionary}

\item{bclose}{biomass that invokes fishing closure}

\item{fmin}{minimum allowable (bycatch) fishing mortality under closure}

\item{obs}{obtion to show observation with input class `FLStock`}

\item{kobe}{add kobe colour-coding}

\item{alpha}{transparency of shading}

\item{xmax}{multiplier for upper default xlim}

\item{ymax}{multiplier for upper default ylim}

\item{ylab}{option customize ylab}

\item{xlab}{option customize xlab}

\item{rel}{option to denote x,y labs as relative B/Btrg and F/Ftrg}

\item{expand}{option to expand the plot area to border - default TRUE}

\item{labels}{annotate reference point labels}

\item{critical}{option to highlight critical zone below blim}

\item{labelslabel.cex=3.}{set size of labels}
}
\value{
ggplot
}
\description{
plotAR
Plots the new proposed ICES advice rule
}
\examples{
data(ple4)
srr = srrTMB(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4))
blim = params(srr)[[2]]
brp = computeFbrp(stock=ple4,sr=srr,proxy="f0.1",blim=blim)
rpt = Fbrp(brp)
plotAR(rpt,btrigger=an(0.8*rpt["Btrg"]))
# Use Bpa as trigger (ICES style)
plotAR(rpt,obs=ple4,bpa=an(1.4*rpt["Blim"]))
# Change kobe to greyscale
plotAR(rpt,obs=ple4,bpa=an(1.4*rpt["Blim"]),kobe=F)
# add fishing closure
plotAR(rpt,obs=ple4,bpa=an(1.4*rpt["Blim"]),kobe=T,bclose=1,fmin=0.02)
# show a relative
plotAR(rpt,obs=ple4,rel=TRUE,bpa=an(1.4*rpt["Blim"]),kobe=T,bclose=1,fmin=0.02)
}
