% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDyns.R
\name{plotbioyr}
\alias{plotbioyr}
\title{plotbioyr()
Plots stock N_a, W_a, M_a and Mat_a across years}
\usage{
plotbioyr(stk, ncol = 2)
}
\arguments{
\item{stk}{stock object class FLStock}

\item{ncol}{number of columns in multiplot}
}
\value{
ggplot
}
\description{
plotbioyr()
Plots stock N_a, W_a, M_a and Mat_a across years
}
\examples{
data(ple4)
plotbioyr(ple4)
}
