% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDyns.R
\name{plotdyn}
\alias{plotdyn}
\title{plotdyn()
Plots stock trajectories at age}
\usage{
plotdyn(stk, ncol = 2)
}
\arguments{
\item{stk}{stock object class FLStock}

\item{ncol}{number of columns in multiplot}
}
\value{
ggplot
}
\description{
plotdyn()
Plots stock trajectories at age
}
\examples{
data(ple4)
plotdyn(ple4)
}
