% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{ploteq}
\alias{ploteq}
\title{ploteq()
Modification of method plot(`FLBRP`) to plot equilibrium output of computeFbrp()}
\usage{
ploteq(
  brps,
  refpts = NULL,
  obs = FALSE,
  dashed = TRUE,
  colours = "missing",
  panels = NULL,
  ncol = 2
)
}
\arguments{
\item{brps}{output object from computeFbrp of class FLBRP}

\item{refpts}{Reference points, defaults are computed refpts from computeFbrp()
\itemize{
  \item Fbrp  
  \item Blim 
  \item B0  
  \item Btri  
}}

\item{obs}{Should observations be plotted? Defaults to `FALSE`.}

\item{dashed}{plots vertical dashed lines to highlight refpts locations}

\item{colours}{refpts colours, default is designed for computeFbrp() output}

\item{panels}{plot panel option 1:4}

\item{ncol}{number of plot panel columns}
}
\value{
ggplot
}
\description{
ploteq()
Modification of method plot(`FLBRP`) to plot equilibrium output of computeFbrp()
}
