% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotFbrps}
\alias{plotFbrps}
\title{plotFbrps()
Modification of method plot(`FLBRP`) to plot equilibrium output of computeFbrp()}
\usage{
plotFbrps(
  brp,
  proxies = c("all", "sprx", "bx"),
  obs = FALSE,
  labels = FALSE,
  shapes = "missing",
  colours = "missing",
  panels = NULL,
  ncol = 2
)
}
\arguments{
\item{brp}{output object from computeFbrp of class FLBRP}

\item{obs}{Should observations be plotted? Defaults to `FALSE`.}

\item{labels}{plot refpts label, default to `FALSE`}

\item{shapes}{refpts symbols, default is designed for computeFbrp() output}

\item{colours}{refpts colours, default is designed for computeFbrp() output}

\item{panels}{plot panel option 1:4}

\item{ncol}{number of plot panel columns}

\item{refpts}{Reference points, defaults are computed refpts from computeFbrp()
\itemize{
  \item all  
  \item sprx 
  \item bx  
}}
}
\value{
ggplot
}
\description{
plotFbrps()
Modification of method plot(`FLBRP`) to plot equilibrium output of computeFbrp()
}
