% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotGFCM}
\alias{plotGFCM}
\title{plotGFCM()}
\usage{
plotGFCM(
  ftgt = 1,
  btgt = 1,
  blim = 0.25,
  btrigger = "missing",
  bthr = 0.5,
  fthr = 1.2,
  bclose = 0,
  fmin = 0,
  fadv = 0.8,
  obs = "missing",
  kobe = TRUE,
  alpha = 1,
  xmax = 1.3,
  ymax = 1.5,
  ylab = "missing",
  xlab = "missing",
  rel = TRUE,
  expand = TRUE,
  labels = TRUE,
  critical = kobe,
  text = TRUE
)
}
\arguments{
\item{btgt}{Btarget corresponding Ftgt}

\item{blim}{biomass limit point used as fraction of Btgt, default 0.25-0.3Btgt}

\item{btrigger}{biomass trigger (can differ from Bthr)}

\item{bthr}{biomass precautionary point as fraction of Bmsy, default 0.5Btgt}

\item{fthr}{upper Ftgt threshold, default 1.2 Ftgt}

\item{bclose}{fishing closure}

\item{fmin}{minimum F at closure}

\item{fadv}{advice F from HCR}

\item{obs}{"missing" option to show observed ssb,fba}

\item{kobe}{if TRUE option to add KOBE colors}

\item{alpha}{option for transparency}

\item{xmax}{multiplier of btgt for ylim}

\item{ymax}{multiplier of ftgt for xlim}

\item{ylab}{}

\item{xlab}{}

\item{rel}{illustrate relative to btgt and ftgt if TRUE}

\item{expand}{= TRUE graphic param}

\item{labels}{if true add labels}

\item{critical}{show area below Blim as dark red}

\item{text}{show text}

\item{ftg}{Ftarget}
}
\value{
GFCM advice plot
}
\description{
Produces candidate GFCM advice plot
}
\examples{
plotGFCM(fadv=0.95,btrigger=0.5)
}
