% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotICES}
\alias{plotICES}
\title{plotICES
Plots the current ICES advice rule}
\usage{
plotICES(
  fmsy = 1,
  blim = 1/1.4,
  btrigger = 1.4,
  bclose = 0,
  fmin = 0,
  obs = "missing",
  kobe = TRUE,
  alpha = 1,
  xmax = 3,
  ymax = 1.5,
  ylab = "missing",
  xlab = "missing",
  rel = FALSE,
  expand = TRUE,
  labels = TRUE,
  critical = FALSE
)
}
\arguments{
\item{fmsy}{Target F = min(Fbrp,Fp0.5)}

\item{blim}{biomass limit}

\item{btrigger}{is also btrg=bthresh}

\item{bclose}{biomass that invokes fishing closure}

\item{fmin}{minimum allowable (bycatch) fishing mortality under closure}

\item{obs}{obtion to show observation with input class `FLStock`}

\item{kobe}{add kobe colour-coding}

\item{alpha}{transparency of shading}

\item{xmax}{multiplier for upper default xlim}

\item{ymax}{multiplier for upper default ylim}

\item{ylab}{option customize ylab}

\item{xlab}{option customize xlab}

\item{rel}{option to denote x,y labs as relative B/Btrg and F/Ftrg}

\item{expand}{option to expand the plot area to border - default TRUE}

\item{labels}{annotate reference point labels}

\item{critical}{option to highlight critical zone below blim}
}
\value{
ggplot
}
\description{
plotICES
Plots the current ICES advice rule
}
\examples{
plotICES()
# Close fishery at Blim and adjust axis labels to relative
plotICES(btrigger=1.4,bclose=1/1.4,rel=TRUE)
# Close fishery at Blim, but allow fmin (e.g. bycatch)
plotICES(btrigger=1.4,bclose=1/1.4,fmin=0.05,rel=TRUE)
# Plot stock data
data(ple4)
plotICES(fmsy=0.25,blim=2e5,,btrigger = 1.4*2e5,obs=ple4)
}
