% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotMajuro}
\alias{plotMajuro}
\title{Plots the new proposed ICES advice rule}
\usage{
plotMajuro(
  ftgt = 1,
  fthresh = 1.1,
  btgt = 1,
  blim = 0.1,
  btrigger = 0.8 * btgt,
  bthresh = 0.5 * btgt,
  bclose = 0,
  fmin = 0,
  obs = "missing",
  kobe = TRUE,
  alpha = 1,
  xmax = 1.5,
  ymax = 1.5,
  ylab = "missing",
  xlab = "missing",
  rel = FALSE,
  expand = TRUE,
  labels = TRUE,
  critical = kobe
)
}
\arguments{
\item{ftgt}{Target F = min(Fbrp,Fp0.5)}

\item{btgt}{Biomass target corresponding to Fbrp}

\item{blim}{biomass limit}

\item{btrigger}{biomass trigger below which F is linearly reduced}

\item{bthresh}{biomass threshold beyond which biomass is classified sustainable}

\item{bclose}{biomass that invokes fishing closure}

\item{fmin}{minimum allowable (bycatch) fishing mortality under closure}

\item{obs}{obtion to show observation with input class `FLStock`}

\item{kobe}{add kobe colour-coding}

\item{alpha}{transparency of shading}

\item{xmax}{multiplier for upper default xlim}

\item{ymax}{multiplier for upper default ylim}

\item{ylab}{option customize ylab}

\item{xlab}{option customize xlab}

\item{rel}{option to denote x,y labs as relative B/Btgt and F/Ftgt}

\item{expand}{option to expand the plot area to border - default TRUE}

\item{labels}{annotate reference point labels}

\item{critical}{option to highlight critical zone below blim}
}
\value{
ggplot
}
\description{
Plots the new proposed ICES advice rule
}
\examples{
plotMajuro()
}
