% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{plotpf}
\alias{plotpf}
\title{plotpf()}
\usage{
plotpf(object, quant = c("vb", "ssb"), fmsy = NULL, rel = FALSE)
}
\arguments{
\item{object}{An *FLBRP*}

\item{quant}{choose between vb and ssb or both}

\item{fmsy}{default if Fmsy}

\item{rel}{if TRUE ratios are produced for spcurve}
}
\value{
ggplot
}
\description{
plots production functions
}
\examples{
data(ple4)
sr <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
brp = FLBRP(ple4,sr)
asem2spm(brp)[1:4]
plotpf(brp)
plotpf(brp,rel=T)
}
