% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDyns.R
\name{plotspr}
\alias{plotspr}
\title{plotspr()
Plots current vs unfished spawning biomass per recruit at age}
\usage{
plotspr(stk, nyears = 3)
}
\arguments{
\item{stk}{stock object class FLStock}

\item{nyears}{number of current last years, default is 3}

\item{ncol}{number of columns in multiplot}
}
\value{
ggplot
}
\description{
plotspr()
Plots current vs unfished spawning biomass per recruit at age
}
\examples{
data(ple4)
plotbioage(ple4)
}
