% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRefs.R
\name{ref.bisect}
\alias{ref.bisect}
\title{Bisection approach to find target F for B (sex-structured)}
\usage{
ref.bisect(
  stock,
  sr,
  deviances = rec(stock) \%=\% 1,
  metrics,
  refpts,
  statistic,
  years,
  pyears = years,
  tune,
  target = 1,
  tol = 0.01,
  maxit = 15,
  verbose = TRUE
)
}
\arguments{
\item{stock}{object class FLStock}

\item{sr}{object class FLSR}

\item{metrics}{FLQuant of FLStock to be defined}

\item{statistic}{}

\item{years}{years to be evaluated}

\item{tune}{range for input x}

\item{target}{target with default 1 for ratios}

\item{tol}{tolerance level}

\item{maxit}{number of optimisation steps}

\item{log}{if TRUE, optimise on log-scale}
}
\description{
The plain bisection algorithm (Burden & Douglas, 1985) is employed here to
find the value of a given forecast target quantity (e.g. `fbar`) for which
a selected value of a performance statistic is obtained over a chosen period.
}
\references{
{Burden, Richard L.; Faires, J. Douglas (1985), "2.1 The Bisection Algorithm", Numerical Analysis (3rd ed.), PWS Publishers, ISBN 0-87150-857-5}
}
\author{
Credits to Iago Mosqueira
}
