% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{rffwd}
\alias{rffwd}
\title{rffwd() Project forward an FLStock with evolutionary Fbar}
\usage{
rffwd(object, sr, fbar = control, control = fbar, deviances = "missing")
}
\arguments{
\item{object}{An *FLStock*}

\item{sr}{A stock-recruit relationship, *FLSR* or *predictModel*.}

\item{fbar}{Yearly target for average fishing mortality, *FLQuant*.}

\item{control}{Yearly target for average fishing mortality, *FLPar*.}

\item{deviances}{Deviances for the strock-recruit relationsip, *FLQuant*.}
}
\value{
The projected *FLStock* object.
}
\description{
rffwd() Project forward an FLStock with evolutionary Fbar
}
\examples{
data(ple4)
sr <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
brp = computeFbrp(ple4,sr,proxy="msy") 
fbar(brp) = FLQuant(rep(0.01,70))
stk = as(brp,"FLStock")
units(stk) = standardUnits(stk)
its = 100
stk <- FLStockR(propagate(stk, its))
stk@refpts= Fbrp(brp)
b0=an(Fbrp(brp)["B0"])
control = FLPar(Feq=0.15,Frate=0.1,Fsigma=0.15,SB0=b0,minyear=2,maxyear=70,its=its)
run <- rffwd(stk, sr=sr,control=control,deviances=ar1rlnorm(0.3, 1:70, its, 0, 0.6))
plotAdvice(run)
}
