% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rGclass}
\alias{rGclass}
\title{Function to characterize Productivity and refpts based on r and Generation}
\usage{
rGclass(r = NULL, gt = NULL)
}
\arguments{
\item{r}{value of the intrinsic rate of population increase}

\item{gt}{generation time G}
}
\value{
list with Productivity category and suggest Fbrps
}
\description{
Function to characterize Productivity and refpts based on r and Generation
}
