% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{sam2FLStockR}
\alias{sam2FLStockR}
\title{sam2FLStockR()}
\usage{
sam2FLStockR(res, itsCI = 1000)
}
\arguments{
\item{res}{fit from FLSAM}

\item{itsCI}{number of iterations to depict uncertainty in plots}
}
\value{
FLStockR with refpts
}
\description{
sam2FLStockR()
}
