% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{sam2stars}
\alias{sam2stars}
\title{sam2stars()}
\usage{
sam2stars(sam, refpts = NULL, quantiles = c(0.05, 0.95))
}
\arguments{
\item{sam}{fit from FLSAM}

\item{refpts}{optional FLPar object with reference points}
}
\value{
STARS list with $timeseris and $refpts
}
\description{
sam2stars()
}
