% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{schaefer.sim}
\alias{schaefer.sim}
\title{schaefer.sim()}
\usage{
schaefer.sim(
  k = 10000,
  r = 0.3,
  q = 0.5,
  pe = 0.1,
  oe = 0.2,
  bk = 0.9,
  years = 1980:2022,
  f0 = 0.2,
  fhi = 2.2,
  flo = 0.8,
  sigmaF = 0.15,
  iters = 1,
  blim = 0.3,
  bthr = 0.5,
  rel = FALSE
)
}
\arguments{
\item{k}{carrying capacity}

\item{r}{intrinsic rate of population increase}

\item{q}{catchability coefficient}

\item{pe}{process error}

\item{oe}{process error}

\item{bk}{initial fraction of b/k}

\item{years}{time horizon}

\item{f0}{factor for initial year as f0 = f/fmsy}

\item{fhi}{factor for high F as fhi = f/fmsy}

\item{flo}{factor for low F as flo = fbar/fmsy}

\item{sigmaF}{variation on f trajectory}

\item{iters}{number of iterations}

\item{rel}{if TRUE metrics B/Bmsy and F/Fmsy are produced}
}
\value{
FLQuants
}
\description{
generates a Schafer surplus production model with process and observation error
}
\examples{
stk = schaefer.sim(iters=100,q=0.5) 
plotAdvice(stk)
plot(FLIndex(index=iter(stk@stock,1))) # index
}
