% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{sops}
\alias{sops}
\title{scales catch-at-age to total catch with error (optional)}
\usage{
sops(object, stock, sigma = 0.1, what = c("catch", "landings", "discards")[1])
}
\arguments{
\item{object}{FLQuant catch.n, discard.n, landings.n}

\item{stock}{FLStock}

\item{sigma}{observation error}

\item{what}{type c("catch", "landings", "discards")}
}
\value{
FLQuant
}
\description{
scales catch-at-age to total catch with error (optional)
}
