% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{spict2FLQuant}
\alias{spict2FLQuant}
\title{spict2FLQuant()}
\usage{
spict2FLQuant(
  x,
  metric = c("ssb", "fbar", "catch", "stock", "harvest")[1],
  forecast = F
)
}
\arguments{
\item{x}{fit from SPICT}

\item{forecast}{TRUE/FALSE}
}
\value{
FLQuant
}
\description{
spict2FLQuant()
}
\author{
adopted from Laurie Kell (biodyn)
}
