% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{spict2FLStockR}
\alias{spict2FLStockR}
\title{spict2FLStockR()}
\usage{
spict2FLStockR(res, bfrac = 0.3, rel = FALSE, forecast = NULL)
}
\arguments{
\item{res}{fit from SPICT}

\item{bfrac}{biomass limit reference point as fraction of Bmsy}

\item{rel}{if TRUE ratios BBmsy and FFmsy are stored}

\item{forecast}{extract forecast TRUE/FALSE}
}
\value{
FLStockR with refpts
}
\description{
spict2FLStockR()
}
