% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{spict2stars}
\alias{spict2stars}
\title{spict2stars()}
\usage{
spict2stars(spict, blim = 0.3, bthr = 0.5, quantiles = c(0.05, 0.95))
}
\arguments{
\item{spict}{fit from  fit.spict()}

\item{blim}{biomass limit point as fraction of Bmsy, default 0.3Bmsy (ICES)}

\item{bthr}{biomass precautionary point as fraction of Bmsy, default 0.5Bmsy (ICES)}
}
\value{
STARS list with $timeseris and $refpts
}
\description{
spict2stars()
}
