% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{ss2FLStockR}
\alias{ss2FLStockR}
\title{ss2FLStockR()}
\usage{
ss2FLStockR(mvln, thin = 10, output = NULL)
}
\arguments{
\item{mvln}{output from ssmvln()}

\item{thin}{thinnig rate of retained iters}

\item{output}{expected outputs presented as "mle" or median of "iters"}
}
\value{
FLStockR with refpts
}
\description{
ss2FLStockR()
}
