% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{ss2stars}
\alias{ss2stars}
\title{ss2stars()}
\usage{
ss2stars(mvln, output = c("iters", "mle")[2], quantiles = c(0.05, 0.95))
}
\arguments{
\item{mvln}{output of ssmvln()}

\item{output}{choice c("iters","mle")[1]}

\item{quantiles}{default is 95CIs as c(0.025,0.975)}
}
\value{
STARS list with $timeseris and $refpts
}
\description{
ss2stars()
}
