% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ss3devs}
\alias{ss3devs}
\title{ss3devs}
\usage{
ss3devs(om, vcv, Fphi = 0.423, bias.correct = TRUE, ...)
}
\arguments{
\item{om}{*FLom* or *FLStock* object}

\item{vcv}{covariance matrix from ss3vcv()}

\item{Fphi}{autocorrelation of F error}

\item{bias.correct}{lognormal bias correction if TRUE}
}
\value{
FLQuants with devs of F and SSB
}
\description{
function to generate MVN assessment error deviations of F and SSB
}
\author{
Henning Winker (GFCM)
}
