% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ss3vcv}
\alias{ss3vcv}
\title{ss3vcv}
\usage{
ss3vcv(ss3rep)
}
\arguments{
\item{ss3rep}{from r4ss::SS_output}
}
\value{
covariance matrix for F and SSB end year
}
\description{
function to generate to extract variance-covariance matrix for F and SSB
}
\author{
Henning Winker (GFCM)
}
