% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssmvln}
\alias{ssmvln}
\title{ssmvln()}
\usage{
ssmvln(
  ss3rep,
  Fref = NULL,
  years = NULL,
  virgin = FALSE,
  mc = 1000,
  weight = 1,
  run = "MVLN",
  addprj = FALSE,
  ymax = NULL,
  xmax = NULL,
  legendcex = 1,
  verbose = TRUE,
  seed = 123,
  observed.catch = FALSE
)
}
\arguments{
\item{ss3rep}{from r4ss::SS_output}

\item{Fref}{Choice of Fratio c("MSY","Btgt","SPR","F01"), correponding to F_MSY and F_Btgt}

\item{years}{single year or vector of years for mvln}

\item{virgin}{if FALSE (default) the B0 base for Bratio is SSB_unfished}

\item{mc}{number of monte-carlo simulations}

\item{weight}{weighting option for model ensembles weight*mc}

\item{run}{qualifier for model run}

\item{addprj}{include forecast years}

\item{ymax}{ylim maximum}

\item{xmax}{xlim maximum}

\item{verbose}{Report progress to R GUI?}

\item{seed}{retains interannual correlation structure like MCMC}

\item{observed.catch}{if FALSE expected catch is used}

\item{out}{choice c("iters","mle")}

\item{plot}{option to show plot}

\item{legendcex=1}{Allows to adjust legend cex}
}
\value{
output list of quant posteriors and mle's
}
\description{
function to generate uncertainty for Stock Synthesis
}
\author{
Henning Winker (GFCM)
}
