% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stock2ratios}
\alias{stock2ratios}
\title{stock2ratios()}
\usage{
stock2ratios(object)
}
\arguments{
\item{object}{of class *FLStockR*}
}
\value{
FLStockR with ratios F/Ftgt and B/Btgt
}
\description{
stock2ratios()
}
