% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stock2spm}
\alias{stock2spm}
\title{stock2spm()}
\usage{
stock2spm(object, brps, bthr = 0.5, blim = 0.3, rel = T, relC = FALSE)
}
\arguments{
\item{object}{of class *FLStockR*}
}
\value{
FLStockR with ratios F/Ftgt and B/Btgt
}
\description{
stock2spm()
}
