% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stockMedians}
\alias{stockMedians}
\title{Converts FLStock into simplified FLStock with Median FLQuants}
\usage{
stockMedians(object, FUN = median)
}
\arguments{
\item{object}{of class *FLStock* or *FLStockR* or *FLStocks*}

\item{FUN}{computes mean, median}
}
\value{
FLStockR with *FLQuants*
}
\description{
Converts FLStock into simplified FLStock with Median FLQuants
}
