% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{updCatch.n}
\alias{updCatch.n}
\title{computes catch.n for a given harvest}
\usage{
updCatch.n(stk)
}
\arguments{
\item{object}{An *FLStock*}
}
\value{
FLQuant
}
\description{
computes catch.n for a given harvest
}
