% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{updsr}
\alias{updsr}
\title{updsr()}
\usage{
updsr(object, s = 0.7)
}
\arguments{
\item{object}{An *FLBRP*}

\item{s}{assumed steepness s}
}
\value{
FLBRP
}
\description{
updates sr in brp after changing biology
}
\examples{
data(ple4)
sr <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
brp = FLBRP(ple4,sr)
s = sr@SV[[1]]
params(brp)
# change
m(brp) = Mlorenzen(stock.wt(brp),Mref=0.15)
brpupd =updsr(brp,s)
params(brp)
}
