% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{ALK}
\alias{ALK}
\title{ALK function}
\usage{
ALK(N_a, iALK)
}
\arguments{
\item{N_a}{numbers at age sample for single event}

\item{iALK}{from iALK() outout}
}
\value{
FLPar of ALK
}
\description{
ALK function
}
