% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lensim.R
\name{alk.sample}
\alias{alk.sample}
\title{generates annual ALK sample with length stratified sampling}
\usage{
alk.sample(lfds, alks, nbin = 20, n.sample = 1)
}
\arguments{
\item{lfds}{length frequency *FLQuant*}

\item{alks}{annual ALK proportions at age output form ALKs() *FLPars*}

\item{nbin}{number of samples per length bin}

\item{n.sample}{sample size of lfd}
}
\value{
FLPars of sampled ALK
}
\description{
generates annual ALK sample with length stratified sampling
}
