% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{ALKs}
\alias{ALKs}
\title{annual ALK function}
\usage{
ALKs(object, iALK)
}
\arguments{
\item{object}{FLQuant with numbers at age}

\item{iALK}{from iALK() outout}
}
\value{
FLPars of ALK
}
\description{
annual ALK function
}
