% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lensim.R
\name{applyALK}
\alias{applyALK}
\title{Convert length frequencies to numbers-at-age}
\usage{
applyALK(lfds, alks)
}
\arguments{
\item{lfds}{An `FLQuant` containing numbers-at-length.}

\item{alks}{An annual `FLPars` collection returned by [ALKs()], or a single
`FLPar` applied to every year.}
}
\value{
An `FLQuant` containing estimated numbers-at-age.
}
\description{
Applies annual age-length keys to length-frequency data to estimate
numbers-at-age by year and iteration.
}
\examples{
\dontrun{
numbers_at_age <- applyALK(lfds, alks)
}

}
