% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{asem2spm}
\alias{asem2spm}
\title{asem2spm()}
\usage{
asem2spm(object, sr)
}
\arguments{
\item{object}{An *FLStock*}

\item{sr}{A stock-recruit relationship, *FLSR* or *predictModel*.}
}
\value{
prior means for r and m *FLPar*
}
\description{
asem2spm()
}
\examples{
data(ple4)
sr <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
asem2spm(ple4,sr)
sr085 <- srrTMB(as.FLSR(ple4,model=bevholtSV),s=0.85,s.est=F,r0=NULL,spr0=mean(spr0y(ple4)))
plotsrs(FLSRs(s.est=sr,s0.85=sr075))
asem2spm(ple4,sr075)
}
