% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{asem2spm}
\alias{asem2spm}
\title{asem2spm()}
\usage{
asem2spm(
  object,
  quant = c("vb", "ssb"),
  fmsy = NULL,
  rel = FALSE,
  spcurve = FALSE
)
}
\arguments{
\item{object}{An *FLBRP*}

\item{quant}{choose between vb and ssb}

\item{rel}{if TRUE ratios are produced for spcurve}

\item{spcurve}{if TRUE a data.frame is added}
}
\value{
prior means for r and m *FLPar*
}
\description{
asem2spm()
}
\examples{
data(ple4)
sr <- srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=mean(spr0y(ple4)))
brp = FLBRP(ple4,sr)
asem2spm(brp)[1:4]
plotpf(brp)
plotpf(brp,rel=T)
}
