% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{ca.sim}
\alias{ca.sim}
\title{generates catch.n with lognormal annual and multinomial age composition observation error}
\usage{
ca.sim(
  object,
  ess = 200,
  what = c("catch", "landings", "discards")[1],
  rescale = FALSE
)
}
\arguments{
\item{object}{FLQuant}

\item{ess}{effective sample size for age composition}

\item{what}{c("catch", "landings", "discards")}

\item{rescale}{rescale to the abudance scale of the input}

\item{sel}{FLQuant with selectivity.pattern e.g. catch.sel()}
}
\value{
FLQuant with catch.n samples
}
\description{
generates catch.n with lognormal annual and multinomial age composition observation error
}
\examples{
data(ple4)
object = propagate(catch.sel(ple4),10)
ca = ca.sim(object,ess=200)
# Checks
ggplot(ca)+geom_line(aes(year,data,col=ac(iter)))+facet_wrap(~age)+
theme(legend.position = "none")+ylab("Index")
}
