% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{ca.sim}
\alias{ca.sim}
\title{generates catch.n with lognormal annual and multinomial age composition observation error}
\usage{
ca.sim(object, ess = 200, what = c("catch", "landings", "discards")[1])
}
\arguments{
\item{object}{FLStock}

\item{ess}{effective sample size for age composition}

\item{what}{c("catch", "landings", "discards")}

\item{sel}{FLQuant with selectivity.pattern e.g. catch.sel()}
}
\value{
FLIndex
}
\description{
generates catch.n with lognormal annual and multinomial age composition observation error
}
\examples{
data(ple4)
object = propagate(ple4,10)
ca = ca.sim(object,ess=200)
# Checks
ggplot(ca)+geom_line(aes(year,data,col=ac(iter)))+facet_wrap(~age)+
theme(legend.position = "none")+ylab("Index")
}
