% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRefs.R
\name{computeFbrp}
\alias{computeFbrp}
\title{computeFbrp()
Computes biological reference points corresponding to the proxy Fbrp}
\usage{
computeFbrp(
  stock,
  sr = NULL,
  proxy = c("sprx", "bx", "f0.1", "msy"),
  x = 40,
  blim = 0.1,
  type = c("b0", "btrg", "value"),
  btri = NULL,
  verbose = T
)
}
\arguments{
\item{stock}{object of class FLStock}

\item{sr}{stock recruitment model of class FLSR}

\item{x}{basis in percent for sprx and bx}

\item{blim}{values < 1 are taken as fraction to B0 and blim > 1 as absolute values unless specified otherwise}

\item{type}{type of blim input, values < 1 are
\itemize{
  \item "b0" fraction to B0  
  \item "btrg" fraction to Btarget    
  \item "value" absolute value
}}

\item{btri}{Btrigger can specified as ratio to Btrg}

\item{verbose}{}

\item{proxies}{choice of Fmsy proxies
\itemize{
  \item "sprx"  spawning potential ratio spr/spr0 with basis x 
  \item "bx" SSB as fraction xSSB0
  \item "f0.1" 10% slope of yield-per-recruit curve
  \item "msy"  maximum surplus production (not defined for segreg)
}}
}
\value{
brp object of class FLBRP with computed Fbrp reference points
}
\description{
computeFbrp()
Computes biological reference points corresponding to the proxy Fbrp
}
