% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeRefs.R
\name{computeFbrps}
\alias{computeFbrps}
\title{computeFbrps()
Computes biological reference points corresponding to the proxy Fbrp}
\usage{
computeFbrps(stock, sr = NULL, proxy = c("all", "sprx", "bx"), verbose = T)
}
\arguments{
\item{stock}{object of class FLStock}

\item{sr}{stock recruitment model of class FLSR}

\item{verbose}{}

\item{proxies}{choice of Fmsy proxies
\itemize{
  \item "all"  both sprx and bx 
  \item "sprx"  spawning potential ratio spr/spr0 with basis x 
  \item "bx" SSB as fraction xSSB0
}}
}
\value{
brp object of class FLBRP with computed Fbrp reference points
}
\description{
computeFbrps()
Computes biological reference points corresponding to the proxy Fbrp
}
