% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRefs.R
\name{Fmmy}
\alias{Fmmy}
\title{Fmmy()
Uses opt.bisect to derive the F at Maximum Median Yield from stochastic simulations}
\usage{
Fmmy(
  brp,
  sigmaR = 0.5,
  rho = 0,
  nyears = 100,
  iters = 250,
  yrs.eval = NULL,
  range = "missing",
  tol = 0.001,
  maxit = 15,
  verbose = TRUE
)
}
\arguments{
\item{brp}{output object from computeFbrp() of class FLBRP}

\item{sigmaR}{lognormal recruitment standard deviation}

\item{rho}{AR1 recruitment autocorrelation coefficient}

\item{nyears}{number of simulation years}

\item{iters}{number simulation iterations}

\item{yrs.eval}{last years to be used evaluation period, default nyears/2}

\item{range}{range of Fbar value to be evaluated}

\item{tol}{tolerance}

\item{maxit}{number of steps}

\item{verbose}{cat comments}
}
\value{
list of  FLPar, FLStock and FLBRP objects
}
\description{
Fmmy()
Uses opt.bisect to derive the F at Maximum Median Yield from stochastic simulations
}
\examples{
data(ple4)
bh = srrTMB(as.FLSR(ple4,model=bevholtSV),spr0=spr0y(ple4))
brp.bh = computeFbrp(ple4,bh,proxy=c("bx","msy"),x=35,blim=0.1)
fmmy.bh = Fmmy(brp.bh,sigmaR=0.7,rho=0.3)
# Compare with deterministic Fmsy
refpts(fmmy.bh$)  
srr = srrTMB(as.FLSR(ple4,model=segreg),spr0=spr0y(ple4),plim=0.15)
brp=computeFbrp(ple4,st,proxy = c("msy","sprx"),x=35,blim=params(st)[[2]])
fmmy(brp,sigmaR=0.7,rho=0.3)
refpts(fmmy.bh$brp)["Fmsy","harvest"]
getF(fmmy.bh) # Stochastic Fmsy
plotFsim(fmmy.bh)
}
