% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRefs.R
\name{Fp05}
\alias{Fp05}
\title{Calculates the Fbar value giving a maximum probability of ssb being below Blim of 5 percent}
\usage{
Fp05(object, iters = "missing", range = "missing", verbose = TRUE)
}
\arguments{
\item{object}{output from Fsim()}

\item{iters}{Number of iterations, cannot exceed input object}

\item{range}{range of Fbar value to be evaluated}

\item{verbose}{Should progress be shown, TRUE.}
}
\value{
Fp.05
}
\description{
Calculates the Fbar value giving a maximum probability of ssb being below Blim of 5 percent
}
