% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRefs.R
\name{Fsim}
\alias{Fsim}
\title{Fsim()
Simulates stochastic stock dynamics under under constant Fbrp}
\usage{
Fsim(
  brp,
  sigmaR = 0.5,
  rho = 0,
  nyears = 100,
  iters = 1000,
  yrs.eval = NULL,
  verbose = TRUE
)
}
\arguments{
\item{brp}{output object from computeFbrp() of class FLBRP}

\item{sigmaR}{lognormal recruitment standard deviation}

\item{rho}{AR1 recruitment autocorrelation coefficient}

\item{nyears}{number of simulation years}

\item{iters}{number simulation iterations}

\item{yrs.eval}{last years to be used evaluation period, default nyears/2}

\item{verbose}{cat comments}
}
\value{
list of  `FLPar`, `FLStock` and `FLBRP` objects
}
\description{
Fsim()
Simulates stochastic stock dynamics under under constant Fbrp
}
