% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{iALK}
\alias{iALK}
\title{inverse ALK function with lmin added to FLCore::invALK}
\usage{
iALK(
  params,
  model = vonbert,
  age,
  cv = 0.1,
  lmin = 5,
  lmax = 1.2,
  bin = 1,
  max = ceiling(linf * lmax),
  reflen = NULL
)
}
\arguments{
\item{params}{growth parameter, default FLPar(linf,k,t0)}

\item{model}{growth model, only option currently vonbert}

\item{age}{age vector}

\item{cv}{of length-at-age}

\item{lmin}{minimum length}

\item{lmax}{maximum upper length specified lmax*linf}

\item{bin}{length bin size, dafault 1}

\item{max}{maximum size value}

\item{reflen}{evokes fixed sd for L_a at sd = cv*reflen}

\item{timing}{t0 assumed 1st January, default seq(0,11/12,1/12), but can be single event 0.5}

\item{unit}{default is "cm"}
}
\value{
FLPar age-length matrix
}
\description{
inverse ALK function with lmin added to FLCore::invALK
}
