% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jb2FLStockR}
\alias{jb2FLStockR}
\title{jb2FLStockR()}
\usage{
jb2FLStockR(jabba, bfrac = 0.3, iters = 2000)
}
\arguments{
\item{jabba}{fit from JABBA fit_jabba() or jabba$kbtrj}
}
\value{
FLStockR with refpts
}
\description{
jb2FLStockR()
}
