% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{len.sim}
\alias{len.sim}
\title{function to generate survey (pulse) and continuous LFDs}
\usage{
len.sim(
  N_a,
  params,
  model = vonbert,
  ess = 250,
  timing = seq(0, 11/12, 1/12),
  unit = "cm",
  scale = TRUE,
  reflen = NULL,
  bin = 1,
  cv = 0.1,
  lmax = 1.2
)
}
\arguments{
\item{N_a}{numbers at age sample}

\item{params}{growth parameter, default FLPar(linf,k,t0)}

\item{model}{growth model, only option currently vonbert}

\item{ess}{effective sample size}

\item{timing}{t0 assumed 1st January, default seq(0,11/12,1/12), but can be single event 0.5}

\item{unit}{default is "cm"}

\item{scale}{if TRUE scaled to N_a input}

\item{reflen}{evokes fixed sd for L_a at sd = cv*reflen}

\item{bin}{length bin size, dafault 1}

\item{cv}{variation in L_a}

\item{lmax}{maximum upper length specified lmax*linf}
}
\value{
FLQuant for length
}
\description{
function to generate survey (pulse) and continuous LFDs
}
