% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMsim.R
\name{lfd.sim}
\alias{lfd.sim}
\title{function to generate survey (pulse) and continuous LFDs}
\usage{
lfd.sim(
  object,
  stock,
  sel = catch.sel(stock),
  params,
  model = vonbert,
  ess = 250,
  timing = seq(0, 11/12, 1/12),
  timeref = 0.5,
  unit = "cm",
  scale = TRUE,
  reflen = NULL,
  bin = 1,
  cv = 0.1,
  lmin = 5,
  lmax = 1.2
)
}
\arguments{
\item{object}{*FLQuant* numbers at age sample}

\item{stock}{*FLStock* object}

\item{sel}{selectivity, default catch.sel(stock)}

\item{params}{growth parameter, default FLPar(linf,k,t0)}

\item{model}{growth model, only option currently vonbert}

\item{ess}{effective sample size}

\item{timing}{default constinoues seq(0,11/12,1/12), but can be single event 0.5}

\item{timeref}{reference timing of the sample, default 0.5 (e.g. survey or catch.n)}

\item{unit}{default is "cm"}

\item{scale}{if TRUE scaled to N_a input}

\item{reflen}{evokes fixed sd for L_a at sd = cv*reflen}

\item{bin}{length bin size, dafault 1}

\item{cv}{variation in L_a}

\item{lmin}{minimum length}

\item{lmax}{maximum upper length specified lmax*linf}
}
\value{
FLQuant for length
}
\description{
function to generate survey (pulse) and continuous LFDs
}
