% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lensim.R
\name{plot_lfd_season}
\alias{plot_lfd_season}
\title{Plot seasonal length-frequency distributions}
\usage{
plot_lfd_season(lfd_season, year, iter = 1)
}
\arguments{
\item{lfd_season}{Named list of seasonal length-frequency `FLQuant` objects.}

\item{year}{Character or numeric year to plot.}

\item{iter}{Integer simulation iteration. Defaults to `1`.}
}
\value{
A `ggplot` object.
}
\description{
Plots seasonal length-frequency distributions for a selected year, with
seasons distinguished by colour and gears shown in separate panels.
}
\examples{
\dontrun{
plot_lfd_season(list(Trawl = lfd_season_trawl, Gillnet = lfd_season_gillnet),
                 year = "2020")
}

}
