% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotAdvice}
\alias{plotAdvice}
\title{plotAdvice
Plots stochastic stock dynamics against refpts for constant Fsim()}
\usage{
plotAdvice(stock, brp, plotrefs = TRUE, ncol = 2, label.size = 3)
}
\arguments{
\item{plotrefs}{if TRUE reference points are plotted}

\item{ncol}{number of plot panel columns}

\item{label.size}{size of refpts labels}

\item{object}{output object from Fsim()}

\item{colour}{color of CIs}
}
\value{
ggplot
}
\description{
plotAdvice
Plots stochastic stock dynamics against refpts for constant Fsim()
}
