% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotAdvice}
\alias{plotAdvice}
\title{plotAdvice
Plots stochastic stock dynamics against refpts for constant Fsim()}
\usage{
plotAdvice(
  stock,
  rpts = "missing",
  plotrefs = TRUE,
  probs = c(0.05, 0.2, 0.5, 0.8, 0.95),
  colour = "dodgerblue",
  ncol = 2,
  label.size = 3
)
}
\arguments{
\item{stock}{FLStock or FLStockR}

\item{plotrefs}{if TRUE reference points are plotted}

\item{probs}{determine credibility intervals, default 80th, 90th percentiles   #' @param ncol number of plot panel columns}

\item{colour}{color of CIs}

\item{label.size}{size of refpts labels}

\item{refpts}{as FLPar or Fbrp() if FLStockR is not provided or should be overwritten}
}
\value{
ggplot
}
\description{
plotAdvice
Plots stochastic stock dynamics against refpts for constant Fsim()
}
\examples{
data(ple4)
srr = srrTMB(as.FLSR(ple4,model=rickerSV),spr0=spr0y(ple4))
brp = computeFbrp(stock=ple4,sr=srr,proxy=c("sprx","f0.1","fe40"),blim=0.1,type="b0")
plotAdvice (ple4,brp)
}
