% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotAdvice}
\alias{plotAdvice}
\title{plotAdvice
Plots stochastic stock dynamics against refpts for constant Fsim()}
\usage{
plotAdvice(stock, rpts = "missing", plotrefs = TRUE, ncol = 2, label.size = 3)
}
\arguments{
\item{stock}{FLStock or FLStockR}

\item{plotrefs}{if TRUE reference points are plotted}

\item{ncol}{number of plot panel columns}

\item{label.size}{size of refpts labels}

\item{refpts}{as FLPar or Fbrp() if FLStockR is not provided or should be overwritten}

\item{colour}{color of CIs}
}
\value{
ggplot
}
\description{
plotAdvice
Plots stochastic stock dynamics against refpts for constant Fsim()
}
\examples{
data(ple4)
srr = srrTMB(as.FLSR(ple4,model=rickerSV),spr0=spr0y(ple4))
brp = computeFbrp(stock=ple4,sr=srr,proxy=c("sprx","f0.1","fe40"),blim=0.1,type="b0")
plotAdvice (ple4,brp)
}
