% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{ploteq}
\alias{ploteq}
\title{ploteq()
Modification of method plot(FLBRP) to plot equilibrium output of computeFbrp()}
\usage{
ploteq(
  brp,
  refpts = c("Fbrp", "Btri", "Blim", "B0"),
  obs = FALSE,
  labels = FALSE,
  shapes = "missing",
  colours = "missing",
  panels = NULL,
  ncol = 2,
  ...
)
}
\arguments{
\item{brp}{output object from computeFbrp of class FLBRP}

\item{refpts}{Reference points, defaults are computed refpts from computeFbrp()
\itemize{
  \item Fbrp Fmsy proxy 
  \item Blim 
  \item B0  unfished biomass for reference period
  \item Btri Btrigger as fraction of Btrg corresponding Fbrp 
}}

\item{obs}{Should observations be plotted? Defaults to `FALSE`.}

\item{labels}{plot refpts label, default to `FALSE`}

\item{shapes}{refpts symbols, default is designed for computeFbrp() output}

\item{colours}{refpts colours, default is designed for computeFbrp() output}

\item{panels}{plot panel option 1:4}

\item{ncol}{number of plot panel columns}
}
\value{
ggplot
}
\description{
ploteq()
Modification of method plot(FLBRP) to plot equilibrium output of computeFbrp()
}
