% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{ploteq}
\alias{ploteq}
\title{ploteq()
Modification of method plot(`FLBRP`) to plot equilibrium output of computeFbrp()}
\usage{
ploteq(
  brps,
  refpts = "missing",
  obs = FALSE,
  rel = FALSE,
  rpf = TRUE,
  dashed = rpf,
  colours = "missing",
  panels = NULL,
  ncol = 2
)
}
\arguments{
\item{brps}{output object from computeFbrp of class FLBRP}

\item{refpts}{Reference points, defaults are computed refpts from computeFbrp()
\itemize{
  \item Fbrp  
  \item Blim 
  \item B0  
  \item Btri  
}}

\item{obs}{Should observations be plotted? Defaults to `FALSE`.}

\item{rel}{option to denote x,y labs as relative B/Btgt and F/Ftgt}

\item{rpf}{adds refpts in plots}

\item{dashed}{plots vertical dashed lines to highlight refpts locations}

\item{colours}{refpts colours, default is designed for computeFbrp() output}

\item{panels}{plot panel option 1:4}

\item{ncol}{number of plot panel columns}
}
\value{
ggplot
}
\description{
ploteq()
Modification of method plot(`FLBRP`) to plot equilibrium output of computeFbrp()
}
\examples{
data(ple4)
srr = srrTMB(as.FLSR(ple4,model=rickerSV),spr0=spr0y(ple4))
brp = computeFbrp(stock=ple4,sr=srr,proxy=c("sprx","f0.1","msy"),blim=0.1,type="b0")
ploteq(brp,obs=TRUE)
ploteq(brp,obs=TRUE,refpts="msy",rel=TRUE)
brp.pa = computeFbrp(stock=ple4,sr=srr,proxy=c("msy","sprx","f0.1"),blim=0.1,bpa=Fbrp(brp)["Blim"]*2,type="b0")
ploteq(brp.pa,obs=TRUE,rel=TRUE)
}
