% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotFsim}
\alias{plotFsim}
\title{plotFsim
Plots stochastic stock dynamics against refpts for constant Fsim()}
\usage{
plotFsim(
  object,
  worms = TRUE,
  thinning = 10,
  probs = c(0.05, 0.2, 0.5, 0.8, 0.95),
  plotrefs = TRUE,
  colour = "dodgerblue",
  ncol = 2,
  yrs.eval = NULL
)
}
\arguments{
\item{object}{output object from Fsim()}

\item{worms}{option to show individual iterations}

\item{thinning}{thinning rate of iterations shows, e.g. 10 shows every 10th}

\item{probs}{determine credibility intervals, default 80th, 90%}

\item{plotrefs}{if TRUE reference points are plotted}

\item{colour}{color of CIs}

\item{ncol}{number of plot panel columns}

\item{yrs.eval}{last years to be used evaluation period, default nyears/2}
}
\value{
ggplot
}
\description{
plotFsim
Plots stochastic stock dynamics against refpts for constant Fsim()
}
