% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRefs.R
\name{plotWKREF}
\alias{plotWKREF}
\title{plotWKREF
Plots the newploiomass target corresponding to Fbrp}
\usage{
plotWKREF(
  ftgt = 1,
  btgt = 1,
  blim = 0.2,
  btrigger = 0.9 * btgt,
  bthresh = 0.8 * btgt,
  bclose = 0,
  fmin = 0,
  obs = "missing",
  kobe = TRUE,
  alpha = 1,
  xmax = 1.3,
  ymax = 1.5,
  ylab = "missing",
  xlab = "missing",
  rel = FALSE,
  expand = TRUE,
  labels = TRUE,
  critical = kobe
)
}
\arguments{
\item{blim}{biomass limit}

\item{btrigger}{biomass trigger below which F is linearly reduced}

\item{bthresh}{biomass threshold beyond which biomass is classified sustainable}

\item{bclose}{ratio biomass/blim that invokes fishing closure relative to blim}

\item{fmin}{minimum allowable (bycatch) fishing mortality under closure}

\item{obs}{obtion to show observation with input class `FLStock`}

\item{kobe}{add kobe colour-coding}

\item{alpha}{transparency of shading}

\item{xmax}{multiplier for upper default xlim}

\item{ymax}{multiplier for upper default ylim}

\item{ylab}{option customize ylab}

\item{xlab}{option customize xlab}

\item{rel}{option to denote x,y labs as relative B/Btgt and F/Ftgt}

\item{expand}{option to expand the plot area to border - default TRUE}

\item{labels}{annotate reference point labels}

\item{critical}{option to highlight critical zone below blim}
}
\value{
ggplot
}
\description{
plotWKREF
Plots the newploiomass target corresponding to Fbrp
}
\examples{
plotWKREF()
# Close fishery at Blim and adjust axis labels to relative
plotWKREF(blim=0.2,bclose=0.2,rel=TRUE)
# Close fishery at Blim, but allow fmin (e.g. bycatch)
plotWKREF(blim=0.2,bclose=0.2,fmin=0.1,rel=TRUE)
# Change Btrigger above Btgt
plotWKREF(blim=0.2,bclose=0.2,fmin=0.1,btrigger=0.80,rel=TRUE)
# Plot stock data
data(ple4)
plotWKREF(ftgt=0.25,btgt=8e+05,btrigger = 0.9*8e+05, blim=2e5,bclose=3e5,fmin=0.03,obs=ple4)
}
