% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{sam2FLQuant}
\alias{sam2FLQuant}
\title{sam2FLQuant()}
\usage{
sam2FLQuant(
  object,
  metric = c("ssb", "fbar", "catch", "rec")[1],
  what = c("mle")
)
}
\arguments{
\item{object}{fit from FLSAM}

\item{forecast}{TRUE/FALSE}
}
\value{
FLQuant
}
\description{
sam2FLQuant()
}
\author{
adopted from Laurie Kell (biodyn)
}
