% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{spict2FLQuant}
\alias{spict2FLQuant}
\title{spict2FLQuant()}
\usage{
spict2FLQuant(
  x,
  metric = c("ssb", "fbar", "catch", "stock", "harvest")[1],
  osa = FALSE,
  forecast = F,
  what = c("mle")
)
}
\arguments{
\item{x}{fit from SPICT}

\item{osa}{add one-step-ahead forecast}

\item{forecast}{TRUE/FALSE}

\item{what}{mle or log.sd}
}
\value{
FLQuant
}
\description{
spict2FLQuant()
}
\author{
adopted from Laurie Kell (biodyn)
}
