% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{spict2FLQuant}
\alias{spict2FLQuant}
\title{spict2FLQuant()}
\usage{
spict2FLQuant(
  x,
  val = c("logB", "logFnotS", "logCpred", "logBBmsy", "logFFmsynotS")[1]
)
}
\arguments{
\item{res}{fit from SPICT}
}
\value{
FLQuant
}
\description{
spict2FLQuant()
}
\author{
adopted from Laurie Kell (biodyn)
}
