% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{spict2FLStockR}
\alias{spict2FLStockR}
\title{spict2FLStockR()}
\usage{
spict2FLStockR(
  res,
  blim = 0.3,
  bthr = 0.5,
  rel = FALSE,
  osa = FALSE,
  forecast = NULL,
  itsCI = 1
)
}
\arguments{
\item{res}{fit from SPICT}

\item{blim}{biomass limit reference point as fraction of Bmsy}

\item{bthr}{biomass precautionary reference point as fraction of Bmsy}

\item{rel}{if TRUE ratios BBmsy and FFmsy are stored}

\item{osa}{add one-step-ahead forecast}

\item{forecast}{extract forecast TRUE/FALSE}
}
\value{
FLStockR with refpts
}
\description{
spict2FLStockR()
}
