% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{ss2FLStockR}
\alias{ss2FLStockR}
\title{ss2FLStockR()}
\usage{
ss2FLStockR(mvln, thin = 10, output = c("iters", "mle")[1])
}
\arguments{
\item{mvln}{output from ssmvln()}

\item{thin}{thinnig rate of retained iters}

\item{output}{choice c("iters","mle")[1]}
}
\value{
FLStockR with refpts
}
\description{
ss2FLStockR()
}
