% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ss2FLStockR}
\alias{ss2FLStockR}
\title{ss2FLStockR()}
\usage{
ss2FLStockR(mvln, iters = 2000, output = c("iters", "mle")[1])
}
\arguments{
\item{mvln}{output from ssmvln()}

\item{iters}{maximum iters retained}

\item{output}{choice c("iters","mle")[1]}
}
\value{
FLStockR with refpts
}
\description{
ss2FLStockR()
}
