% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stock2ratios}
\alias{stock2ratios}
\title{spict2stock2ratios()}
\usage{
stock2ratios(object, bfrac = 0.3, rel = FALSE)
}
\arguments{
\item{object}{of class *FLStockR*}

\item{bfrac}{biomass limit reference point as fraction of Bmsy}
}
\value{
FLStockR with ratios F/Ftgt and B/Btgt
}
\description{
spict2stock2ratios()
}
