% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stockMedians}
\alias{stockMedians}
\title{stockMedians}
\usage{
stockMedians(stock)
}
\arguments{
\item{stock}{class FLStock or FLStockR}
}
\value{
medians of all FLstock FlQuants
}
\description{
stockMedians
}
