% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{updstars}
\alias{updstars}
\title{updstars()}
\usage{
updstars(star, newrefpts)
}
\arguments{
\item{star}{output of star list with new refpoints}

\item{newrefpts}{FLPar manually adjusted reference points}
}
\value{
STARS list with $timeseris and $refpts
}
\description{
updstars()
}
